+++ tests/enums/ConfigNames.php
@@ -6,6 +6,7 @@ namespace Test\enums;
 
 class ConfigNames
 {
+    public const DIADOC_AUTH_MODE = 'DIADOC_AUTH_MODE';
     public const OAUTH_CLIENT_ID = 'OAUTH_CLIENT_ID';
     public const OAUTH_CLIENT_SECRET = 'OAUTH_CLIENT_SECRET';
     public const OAUTH_IDENTITY_URL = 'OAUTH_IDENTITY_URL';
@@ -16,6 +17,11 @@ class ConfigNames
     public const DIADOC_REFRESH_TOKEN = 'DIADOC_REFRESH_TOKEN';
     public const DIADOC_ACCESS_EXPIRES_AT = 'DIADOC_ACCESS_EXPIRES_AT';
     public const DIADOC_OAUTH_CACHE_KEY = 'diadoc_oauth_session';
+    public const DD_AUTH = 'DD_AUTH';
+    public const AUTH_LOGIN = 'AUTH_LOGIN';
+    public const AUTH_PASSWORD = 'AUTH_PASSWORD';
+    public const DIADOC_LEGACY_TOKEN = 'DIADOC_LEGACY_TOKEN';
+    public const DIADOC_LEGACY_CACHE_KEY = 'diadoc_legacy_auth_token';
     public const ORG_ID = 'ORG_ID';
     public const FROM_BOX_ID = 'FROM_BOX_ID';
     public const TO_BOX_ID = 'TO_BOX_ID';